% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PI_calculation.R
\name{calculate_inflection_points}
\alias{calculate_inflection_points}
\title{Calculate Inflection Point (IPi) for PI matrix}
\usage{
calculate_inflection_points(data, threshold = 0.5)
}
\arguments{
\item{data}{A data frame containing sample_class and ENST columns with output from function 'PI_matrix'.}

\item{threshold}{A numeric threshold for determining whether the value in ENST columns is counted (default is 0.5).}
}
\value{
A data frame with proportions for each tissue and transcript (ENST columns).
}
\description{
This function calculates the Inflection point which is a point on the PDF where the double-differential = 0.
than or equal to a threshold (iter) for each transcript (ENST columns) grouped by sample class,
then merging with tissue data and converting counts to proportions.
}
