% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PI_calculation.R
\name{divide_transtex_classes}
\alias{divide_transtex_classes}
\title{Process Data into Expression Classes}
\usage{
divide_transtex_classes(data)
}
\arguments{
\item{data}{A list which is the output after applying Inflection point cutoff.}
}
\value{
A list containing the categorized data frames and a count table.
}
\description{
This function classifies transcripts into different expression categories such as tissue-specific,
tissue-enhanced, widespread based on their expression in a number of tissues.
}
