% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fold_change.R
\name{fold_change_max_TSp}
\alias{fold_change_max_TSp}
\title{Calculate Fold Change Based on Maximum Value for Tissue-Specific Transcripts}
\usage{
fold_change_max_TSp(data)
}
\arguments{
\item{data}{A matrix with tissues, samples and transcript/genes as columns (expression matrix)}
}
\value{
A data frame with tissues and log2-transformed fold changes.
}
\description{
This function calculates the fold change for each transcript in each tissue
compared to its maximum expression across the rest of tissues tissues, then applies a log2 transformation.
}
