% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fold_change.R
\name{fold_change_TEn}
\alias{fold_change_TEn}
\title{Calculate Fold Change for Tissue-Enhanced Transcripts}
\usage{
fold_change_TEn(df1, df2)
}
\arguments{
\item{df1}{Data frame with mean TPM values (mTPM) for each sample class, and additional metadata.}

\item{df2}{Data frame with PI tissue groupings for tissue-enhanced calculations.}
}
\value{
A modified data frame and log2 fold change calculations.
}
\description{
Calculate Fold Change for Tissue-Enhanced Transcripts
}
