% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fold_change.R
\name{fold_change_TSp}
\alias{fold_change_TSp}
\title{Calculate Fold Change for Tissue-Specific}
\usage{
fold_change_TSp(data, category = "tissue-specific", tissue_interest)
}
\arguments{
\item{data}{A data frame containing TPM values, sample classes, and other relevant metadata.}

\item{category}{A character string specifying the analysis category: 'tissue-specific' or 'tissue-enhanced'.}

\item{tissue_interest}{A character vector of tissue names of interest for tissue-specific calculations.}

\item{df2}{A data frame containing PI tissue groupings for tissue-enhanced calculations.}
}
\value{
A data frame with log2 fold change calculations.
}
\description{
Calculate Fold Change for Tissue-Specific
}
