% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PI_calculation.R
\name{prop_greater}
\alias{prop_greater}
\title{Calculate Inflection Point (IPi) helper function to determine transcripts above IPi}
\usage{
prop_greater(data_prop, up_lim)
}
\arguments{
\item{data_prop}{PI matrix.}

\item{up_lim}{cutoff threshold.}
}
\value{
A list containing the filtered values above threshold.
}
\description{
Calculate Inflection Point (IPi) helper function to determine transcripts above IPi
}
