% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample.R
\name{subsampling_empiricalP}
\alias{subsampling_empiricalP}
\title{Perform Stratified Sampling, Hypothesis Testing}
\usage{
subsampling_empiricalP(
  data,
  tissue_of_interest,
  size,
  iterations = 100,
  threshold = 0.05
)
}
\arguments{
\item{data}{A data frame containing the dataset for analysis with 'tissue' and 'ENST' prefixed columns for transcripts.}

\item{tissue_of_interest}{The tissue of interest for comparison.}

\item{size}{The desired sample size or proportion of the total, passed directly to `sample_size_calculator`.}

\item{iterations}{The number of iterations to perform the subsampling and testing.}

\item{threshold}{The p-value threshold used for empirical p-value calculation.}
}
\value{
A data frame with transcript IDs, empirical p-values, and other metadata if specified.
}
\description{
This function integrates the process of calculating sample sizes, performing stratified sampling and hypothesis testing,
and finally calculating empirical p-values for tissue-specific analysis.
}
