% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fold_change.R
\name{TenhTrans_FC_max}
\alias{TenhTrans_FC_max}
\title{Calculate Fold Change for Tissue-Enhanced Transcripts Based on Maximum Value Outside Tissue Group}
\usage{
TenhTrans_FC_max(df1, df2)
}
\arguments{
\item{df1}{A data frame with TPM values, sample class, and possibly other metadata.
It should include a 'num' column for weighting in mean calculations.}

\item{df2}{A data frame with PI tissue groupings for tissue-enhanced calculations.}
}
\value{
A modified df2 with additional columns for combined mean, maximum outside mean, and log2 fold change.
}
\description{
This function computes the mean TPM for specified groups of tissues, compares it to the maximum mean TPM
outside those groups, and calculates the log2 fold change max.
}
